git-svn-id: http://svn.codergrid.de/svn/sf/bruteripper/trunk@2100 7add9c9f-332d-0410-adcc-f7b1da7091d4

+++ src/go/ffmpeg/Encoder.go
@@ -24,10 +24,10 @@ func(c * Encoder)Open(){
 }
 
 func(c * Encoder)Encode(f * Frame)*Packet{
-  if(c.Ctx.ctx.codec_type==CODEC_TYPE_VIDEO){
+  if(int32(c.Ctx.ctx.codec_type)==CODEC_TYPE_VIDEO){
     return c.encodeVideo(f)
   }
-  if(c.Ctx.ctx.codec_type==CODEC_TYPE_AUDIO){
+  if(int32(c.Ctx.ctx.codec_type)==CODEC_TYPE_AUDIO){
     return c.encodeAudio(f)
   }
   return nil
